Factoring out the core of `DevicesAndImagesDialog`
This logic is needed to function in the upcoming Machine Configuration dialog
