+++ src/models/mod.rs
@@ -1,2 +1,3 @@
 pub mod collectionsview;
+pub mod devimages;
 pub mod itemstable;
