+++ ui/devimages.slint
@@ -1,12 +1,6 @@
 import { Button, VerticalBox, HorizontalBox, ComboBox, ListView, LineEdit } from "std-widgets.slint";
+import { DevicesAndImagesList, DeviceAndImageEntry, DevicesAndImagesState } from "devimageslist.slint";
 
-export struct DeviceAndImageEntry {
-    indent: int,
-    display-tag: string,
-    option-names: [string],
-    option-descriptions: [string],
-    current-option-index: int,
-    filename: string}
 
 export component DevicesAndImagesDialog inherits Window {
     title: "Devices And Images";
@@ -15,56 +9,28 @@ export component DevicesAndImagesDialog inherits Window {
     callback apply-changes-clicked();
     preferred-width: 600px;
     preferred-height: 500px;
-    in property <length> splitter: 160px;
-    in property <[DeviceAndImageEntry]> entries;
-    in property <string> none-string;
+    in property <DevicesAndImagesState> state;
     in property <bool> config-dirty;
     callback entry-option-changed(int, string);
     callback entry-button-clicked(int, Point);
     callback menu-entry-activated(MenuEntry);
     public function show-context-menu(entries: [MenuEntry], point: Point) {
-        context-menu.show(entries, point);
+        dev-images-list.show-context-menu(entries, point);
     }
+
+    // controls
     VerticalBox {
-        ListView {
-            for data[index] in root.entries: Rectangle {
-                height: 30px;
-                width: parent.width;
-                Text {
-                    x: data.indent * 20px;
-                    text: data.display_tag;
-                }
-
-                ComboBox {
-                    x: root.splitter;
-                    width: root.width - root.splitter - 10px;
-                    visible: data.current-option-index >= 0;
-                    model: data.option-descriptions;
-                    current-index: data.current-option-index;
-                    selected(value) => {
-                        root.entry-option-changed(index, data.option-names[self.current-index]);
-                    }
-                }
-
-                LineEdit {
-                    x: root.splitter;
-                    width: root.width - root.splitter - 50px;
-                    visible: data.current-option-index < 0;
-                    placeholder-text: root.none-string;
-                    text: data.filename;
-                    read-only: true;
-                }
-
-                Button {
-                    x: root.width - 50px;
-                    width: 30px;
-                    visible: data.current-option-index < 0;
-                    enabled: !root.config-dirty;
-                    text: "...";
-                    clicked => {
-                        root.entry-button-clicked(index, { x: self.absolute-position.x, y: self.absolute-position.y + self.height });
-                    }
-                }
+        dev-images-list := DevicesAndImagesList {
+            state: root.state;
+            images-button-enabled: !config-dirty;
+            entry-option-changed(index, option-name) => {
+                root.entry-option-changed(index, option-name);
+            }
+            entry-button-clicked(index, point) => {
+                root.entry-button-clicked(index, point);
+            }
+            menu-entry-activated(entry) => {
+                root.menu-entry-activated(entry);
             }
         }
 
@@ -94,12 +60,6 @@ export component DevicesAndImagesDialog inherits Window {
                     root.ok-clicked();
                 }
             }
-        }
-    }
-
-    context-menu := ContextMenu {
-        activated(entry) => {
-            menu-entry-activated(entry)
         }
     }
 }
