+++ ui/main.slint
@@ -5,7 +5,8 @@ import { PathsDialog } from "paths.slint";
 import { MessageBoxDialog } from "messagebox.slint";
 import { NameCollectionDialog } from "namecollection.slint";
 import { ConnectToSocketDialog } from "socket.slint";
-import { DevicesAndImagesDialog, DeviceAndImageEntry } from "devimages.slint";
+import { DevicesAndImagesDialog } from "devimages.slint";
+import { DeviceAndImageEntry } from "devimageslist.slint";
 import { AppWindow } from "appwindow.slint";
 
 export { AboutDialog, AppWindow, ConnectToSocketDialog, LoadingDialog, MessageBoxDialog, NameCollectionDialog, PathsDialog, DevicesAndImagesDialog, DeviceAndImageEntry, Icons }
